Codigo/BackEnd/requirements.txt
# ========== CORE API ==========
fastapi>=0.104.0
uvicorn==0.34.3
gunicorn==21.2.0
python-multipart==0.0.9
requests==2.31.0
python-dotenv==1.0.0
# ========== LLM ==========
groq>=0.8.0
# ========== BANCO DE DADOS VETORIAL ==========
pinecone-client==3.2.1
sentence-transformers==2.7.0
# Torch 2.2.x in this stack is stable with NumPy 1.26
numpy>=1.26.0,<2.0.0
# ========== LANGCHAIN PARA CHUNKING INTELIGENTE ==========
langchain-text-splitters>=0.2.0
langchain-core>=0.2.0
# ========== PROCESSAMENTO DE DOCUMENTOS ==========
pypdf==4.0.1
# ========== AUTENTICACAO ==========
passlib[bcrypt]>=1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
# ========== VALIDACAO E UTILITARIOS ==========
email-validator>=2.0.0
python-dateutil>=2.8.2
# ========== DEPENDENCIAS BASICAS ==========
pydantic>=2.0.0,<3.0.0
typing-extensions>=4.9.0
# ========== GOOGLE OAUTH ==========
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
